Merge uid/gid tables in superblock - now there's a total of 65536 total possible uids/gids, they have to be handled completely differently.
Also add a compression type field to the super_block.  Now there's
LZO compression in the kernel and possibly in the future LZMA will be
supported, need to indicate which compression is being used - without
all the current mess in the third-party LZMA ports.

Lastly try to future proof the layout by adding nominal support for an
ACL/XATTR lookup table to the super_block and an opaque xattr field
to the lreg inode type.