stable: disable action code support by default
The actions code is unfinished and is still "experimental".
As such it is not considered suitable for a stable version of the
squashfs-tools, so by default disable it.

If it is really wanted, it can still be enabled by uncommenting the
ENABLE_ACTIONS = 1
line in the Makefile.

The actions code has been disabled by the straightforward way of
wrapping the code in #defines, and in the hook functions to the
rest of mksquashfs, arranging the #defines so the functions simply
return the default no actions results.  This is a slightly clumsy
way of disabling the code, but, it is easy to verify the resultant
code is correct, and any more "elegant" way of disabling the code
runs the risk of inadvertantly introducing bugs, which defeats the
purpose of adding this code to the stable branch.  So the clumsy
approach is considered by far the best approach.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>